programmers-intellij/src/test/java/kakaopay_5/SolutionTest.java
package kakaopay_5;

import org.junit.Test;

import static org.junit.Assert.*;
import static org.assertj.core.api.Assertions.assertThat;

public class SolutionTest {
    Solution problem = new Solution();

    @Test
    public void test() {
        int[][] land = {{1, 4, 8, 10}, {5, 5, 5, 5}, {10, 10, 10, 10}, {10, 10, 10, 20}};
        int height = 3;
        assertThat(problem.solution(land, height)).isEqualTo(15);
    }

    @Test
    public void test2() {
        int[][] land = {{10, 11, 10, 11}, {2, 21, 20, 10}, {1, 20, 21, 11}, {2, 1, 2, 1}};
        int height = 1;
        assertThat(problem.solution(land, height)).isEqualTo(18);
    }
}